Add basic styling to Video page. Fixes #289
