+++ web/src/hooks/useTvGuide.ts
@@ -1,26 +1,38 @@
-import { QueryClient, UseQueryOptions, useQuery } from '@tanstack/react-query';
+import {
+  QueryClient,
+  UseQueryOptions,
+  UseQueryResult,
+  useQuery,
+  useQueryClient,
+} from '@tanstack/react-query';
 import { ChannelLineup } from '@tunarr/types';
 import { Dayjs } from 'dayjs';
 import { apiClient } from '../external/api.ts';
+import { identity, isUndefined } from 'lodash-es';
 
 const dateRangeQueryKey = (range: { from: Dayjs; to: Dayjs }) =>
   `${range.from.unix()}_${range.to.unix()}`;
 
-const lineupQueryOpts = (
+function lineupQueryOpts<Out = ChannelLineup | undefined>(
   channelId: string,
   range: { from: Dayjs; to: Dayjs },
-) => ({
-  queryKey: ['channels', channelId, 'guide', dateRangeQueryKey(range)],
-  queryFn: async () => {
-    return apiClient.get('/api/channels/:id/lineup', {
-      params: { id: channelId },
-      queries: {
-        from: range.from.toISOString(),
-        to: range.to.toISOString(),
-      },
-    });
-  },
-});
+  mapper: (lineup: ChannelLineup | undefined) => Out = identity,
+) {
+  return {
+    queryKey: ['channels', channelId, 'guide', dateRangeQueryKey(range)],
+    queryFn: async () => {
+      return apiClient
+        .get('/api/channels/:id/lineup', {
+          params: { id: channelId },
+          queries: {
+            from: range.from.toISOString(),
+            to: range.to.toISOString(),
+          },
+        })
+        .then(mapper);
+    },
+  };
+}
 
 const allLineupsQueryOpts = (range: {
   from: Dayjs;
@@ -45,6 +57,49 @@ export const useTvGuide = (params: {
   useQuery(
     lineupQueryOpts(params.channelId, { from: params.from, to: params.to }),
   );
+
+export const useTvGuides = (
+  channelId: string,
+  params: { from: Dayjs; to: Dayjs },
+  extraOpts: Partial<UseQueryOptions<ChannelLineup[]>> = {},
+): UseQueryResult<ChannelLineup[], Error> => {
+  const singleChannelResult = useQuery({
+    ...lineupQueryOpts(channelId, params, (lineup) =>
+      !isUndefined(lineup) ? [lineup] : [],
+    ),
+    ...extraOpts,
+    enabled: channelId !== 'all',
+  });
+
+  const allChannelsResult = useAllTvGuides(params, {
+    ...extraOpts,
+    enabled: channelId === 'all',
+  });
+
+  return channelId === 'all' ? allChannelsResult : singleChannelResult;
+};
+
+export const useTvGuidesPrefetch = (
+  channelId: string,
+  params: { from: Dayjs; to: Dayjs },
+  extraOpts: Partial<UseQueryOptions<ChannelLineup[]>> = {},
+) => {
+  const queryClient = useQueryClient();
+  const query: UseQueryOptions<ChannelLineup[]> =
+    channelId !== 'all'
+      ? {
+          ...lineupQueryOpts(channelId, params, (lineup) =>
+            !isUndefined(lineup) ? [lineup] : [],
+          ),
+          ...extraOpts,
+        }
+      : {
+          ...allLineupsQueryOpts(params),
+          ...extraOpts,
+        };
+
+  queryClient.prefetchQuery(query).catch(console.error);
+};
 
 export const useAllTvGuides = (
   params: { from: Dayjs; to: Dayjs },
