+++ web/src/router.tsx
@@ -58,11 +58,11 @@ export const router = createBrowserRouter(
           errorElement: <ErrorPage />,
         },
         {
-          element: <GuidePage />,
+          element: <GuidePage channelId="all" />,
           index: true,
         },
         {
-          element: <GuidePage />,
+          element: <GuidePage channelId="all" />,
           path: '/guide',
         },
         {
@@ -109,7 +109,7 @@ export const router = createBrowserRouter(
         },
         {
           path: '/guide',
-          element: <GuidePage />,
+          element: <GuidePage channelId="all" />,
         },
         {
           path: '/watch',
